Utvider piloten til 10% av alle virksomheter, og endrer tekster i henhold til ny utarbeidelse
