+++ src/main/kotlin/no/nav/helse/inntektsmeldingsvarsel/dependencyinjection/KoinProfiles.kt
@@ -107,208 +107,6 @@ fun buildAndTestConfig() = module {
 
 }
 
-fun localDevConfig(config: ApplicationConfig) = module {
-    single { getDataSource(createLocalHikariConfig(), "im-varsel", null) as DataSource }
-
-    single {
-        VarslingsmeldingKafkaClient(mutableMapOf<String, Any>(
-                "bootstrap.servers" to "localhost:9092",
-                "max.poll.interval.ms" to "30000")
-                , config.getString("altinn_melding.kafka_topic")) as ManglendeInntektsmeldingMeldingProvider
-    }
-
-    single { VarslingMapper(get()) }
-
-    single { object : RestStsClient { override fun getOidcToken(): String { return "fake token"} } as RestStsClient }
-    single {MockReadReceiptProvider() as ReadReceiptProvider}
-
-    single { object: PdlClient {
-            override fun person(ident: String): PdlPerson? = PdlPerson(listOf(PdlPersonNavn("Test", null, "Testesen")))
-        } as PdlClient
-    }
-
-    single { PostgresVarslingRepository(get()) as VarslingRepository }
-    single { PostgresVentendeBehandlingerRepository(get()) as VentendeBehandlingerRepository }
-    single { VarslingService(get(), get(), get(), get(), get(), get(), AllowAll()) }
-
-    single { MockVarslingSender(get()) as VarslingSender }
-    single { PollForVarslingsmeldingJob(get(), get()) }
-    single { SendVarslingJob(get(), get()) }
-    single { UpdateReadStatusJob(get(), get()) }
-
-    single { PostgresAltinnBrevUtsendelseRepository(get()) as AltinnBrevUtsendelseRepository }
-    single { PostgresAltinnBrevmalRepository(get(), get()) as AltinnBrevMalRepository }
-    single { MockAltinnBrevutsendelseSender() as AltinnBrevutsendelseSender }
-    single { SendAltinnBrevUtsendelseJob(get(), get()) }
-}
-
-@KtorExperimentalAPI
-fun preprodConfig(config: ApplicationConfig) = module {
-    single {
-        getDataSource(createHikariConfig(config.getjdbcUrlFromProperties()),
-                config.getString("database.name"),
-                config.getString("database.vault.mountpath")) as DataSource
-    }
-
-    single {
-        VarslingsmeldingKafkaClient(mutableMapOf(
-                "bootstrap.servers" to config.getString("kafka.endpoint"),
-                CommonClientConfigs.SECURITY_PROTOCOL_CONFIG to "SASL_SSL",
-                SaslConfigs.SASL_MECHANISM to "PLAIN",
-                SaslConfigs.SASL_JAAS_CONFIG to "org.apache.kafka.common.security.plain.PlainLoginModule required " +
-                        "username=\"${config.getString("kafka.username")}\" password=\"${config.getString("kafka.password")}\";"
-        ), config.getString("altinn_melding.kafka_topic")) as ManglendeInntektsmeldingMeldingProvider
-    }
-
-    single {
-        val altinnMeldingWsClient = Clients.iCorrespondenceExternalBasic(
-                config.getString("altinn_melding.pep_gw_endpoint")
-        )
-
-        val sts = wsStsClient(
-                config.getString("sts_url"),
-                config.getString("service_user.username") to config.getString("service_user.password")
-        )
-
-        sts.configureFor(altinnMeldingWsClient)
-
-        altinnMeldingWsClient
-    }
-
-    single { VarslingMapper(get()) }
-
-/*    single {
-        AltinnVarselSender(
-                get(),
-                AltinnVarselMapper(config.getString("altinn_melding.service_id")),
-                get(),
-                get(),
-                config.getString("altinn_melding.username"),
-                config.getString("altinn_melding.password")
-        ) as VarslingSender
-    }*/
-
-    single { MockVarslingSender(get()) as VarslingSender }
-
-    single { DokarkivKlientImpl(config.getString("dokarkiv.base_url"), get(), get()) as DokarkivKlient }
-    single {
-        AltinnReadReceiptClient(
-                get(),
-                config.getString("altinn_melding.username"),
-                config.getString("altinn_melding.password"),
-                config.getString("altinn_melding.service_id"),
-                get()
-        ) as ReadReceiptProvider
-    }
-
-    single { PostgresVarslingRepository(get()) as VarslingRepository }
-    single { PostgresVentendeBehandlingerRepository(get()) as VentendeBehandlingerRepository }
-    single { RestStsClientImpl(config.getString("service_user.username"), config.getString("service_user.password"), config.getString("sts_rest_url"), get()) as RestStsClient }
-    single { PdlClientImpl(config.getString("pdl_url"), get(), get(), get()) as PdlClient }
-
-    single { VarslingService(get(), get(), get(), get(), get(), get(), AllowAll()) }
-
-    single { PollForVarslingsmeldingJob(get(), get()) }
-    single { SendVarslingJob(get(), get()) }
-    single { UpdateReadStatusJob(get(), get())}
-
-    single { PostgresAltinnBrevUtsendelseRepository(get()) as AltinnBrevUtsendelseRepository }
-    single { PostgresAltinnBrevmalRepository(get(), get()) as AltinnBrevMalRepository }
-    single { AltinnBrevutsendelseSenderImpl(get(), get(), get(),
-                config.getString("altinn_melding.username"),
-                config.getString("altinn_melding.password")
-            ) as AltinnBrevutsendelseSender
-    }
-    single { SendAltinnBrevUtsendelseJob(get(), get()) }
-}
-
-@KtorExperimentalAPI
-fun prodConfig(config: ApplicationConfig) = module {
-    single {
-        getDataSource(createHikariConfig(config.getjdbcUrlFromProperties()),
-                config.getString("database.name"),
-                config.getString("database.vault.mountpath")) as DataSource
-    }
-
-    single {
-        VarslingsmeldingKafkaClient(mutableMapOf(
-                "bootstrap.servers" to config.getString("kafka.endpoint"),
-                CommonClientConfigs.SECURITY_PROTOCOL_CONFIG to "SASL_SSL",
-                SaslConfigs.SASL_MECHANISM to "PLAIN",
-                SaslConfigs.SASL_JAAS_CONFIG to "org.apache.kafka.common.security.plain.PlainLoginModule required " +
-                        "username=\"${config.getString("kafka.username")}\" password=\"${config.getString("kafka.password")}\";"
-        ), config.getString("altinn_melding.kafka_topic")) as ManglendeInntektsmeldingMeldingProvider
-    }
-
-    single {
-        val altinnMeldingWsClient = Clients.iCorrespondenceExternalBasic(
-                config.getString("altinn_melding.pep_gw_endpoint")
-        )
-
-        val client = ClientProxy.getClient(altinnMeldingWsClient)
-        //client.inInterceptors.add(LoggingInInterceptor())
-        //client.outInterceptors.add(LoggingOutInterceptor())
-
-        val sts = wsStsClient(
-                config.getString("sts_url"),
-                config.getString("service_user.username") to config.getString("service_user.password")
-        )
-
-        sts.configureFor(altinnMeldingWsClient)
-
-        altinnMeldingWsClient
-    }
-
-    single { PostgresVarslingRepository(get()) as VarslingRepository }
-    single { PostgresVentendeBehandlingerRepository(get()) as VentendeBehandlingerRepository }
-
-    single { VarslingMapper(get()) }
-    single { DokarkivKlientImpl(config.getString("dokarkiv.base_url"), get(), get()) as DokarkivKlient }
-
-    single {
-        AltinnVarselSender(
-                get(),
-                AltinnVarselMapper(config.getString("altinn_melding.service_id")),
-                get(),
-                get(),
-                config.getString("altinn_melding.username"),
-                config.getString("altinn_melding.password")
-        ) as VarslingSender
-    }
-
-    single { RestStsClientImpl(config.getString("service_user.username"), config.getString("service_user.password"), config.getString("sts_rest_url"), get()) as RestStsClient }
-    single { PdlClientImpl(config.getString("pdl_url"), get(), get(), get() ) as PdlClient}
-
-    single { VarslingService(get(), get(), get(), get(), get(), get(), ResourceFileAllowList("/allow-list/virksomheter-allow-prod")) }
-    single {
-        AltinnReadReceiptClient(
-                get(),
-                config.getString("altinn_melding.username"),
-                config.getString("altinn_melding.password"),
-                config.getString("altinn_melding.service_id"),
-                get()
-        ) as ReadReceiptProvider
-    }
-
-    single { PollForVarslingsmeldingJob(get(), get()) }
-    single { SendVarslingJob(get(), get()) }
-
-    single { UpdateReadStatusJob(get(), get()) }
-
-
-    single { PostgresAltinnBrevUtsendelseRepository(get()) as AltinnBrevUtsendelseRepository }
-    single { PostgresAltinnBrevmalRepository(get(), get()) as AltinnBrevMalRepository }
-
-    single { AltinnBrevutsendelseSenderImpl(get(), get(), get(),
-            config.getString("altinn_melding.username"),
-            config.getString("altinn_melding.password")
-    ) as AltinnBrevutsendelseSender
-    }
-    single { SendAltinnBrevUtsendelseJob(get(), get()) }
-
-
-}
-
 // utils
 @KtorExperimentalAPI
 fun ApplicationConfig.getString(path: String): String {
